Add files via upload
The first draft of the script to create the database, along with a file detailing all the operations that we will be performing on the database. This is the first draft, and thus is very likely to be an incomplete list of operations and tables that will be present in the final draft. Operations can also be optimized at a later time by adding subtables and indexes to the database tables.